Fix issue with rewound response bodies
When middleware returns an `HtmlResponse`, the body stream is typically
rewound, with the pointer at position 0. Unfortunately, this means that
any `write()` operations will *overwrite*!

This patch does a check to see if the response is at `eof()`, and, if
not and the response is seekable, sets the pointer to the end of the
stream prior to writing.